updates on gaussian_process_torch: expanded the hyperparameter search to find potentially more accurate emulators
