+++ autoemulate/emulators/gaussian_processs_utils/__init__.py
@@ -1,0 +1,3 @@
+from .polynomial_features import PolynomialFeatures
+from .poly_mean import PolyMean
+from .early_stopping_criterion import EarlyStoppingMax
