+++ autoemulate/emulators/gaussian_processs_utils/early_stopping_criterion.py
@@ -1,0 +1,17 @@
+import numpy as np
+
+from skorch.callbacks import EarlyStopping
+
+class EarlyStoppingMax(EarlyStopping):
+    def _calc_new_threshold(self, score):
+        """Determine threshold based on score."""
+        if self.threshold_mode == 'rel':
+            abs_threshold_change = self.threshold * np.abs(score)
+        else:
+            abs_threshold_change = self.threshold
+
+        if self.lower_is_better:
+            new_threshold = score - abs_threshold_change
+        else:
+            new_threshold = score + abs_threshold_change
+        return new_threshold
